feat: move fs helpers to separate file
